refactor: 로그인 화면 방식 변경
